Вынес типы для спрайтов и раскладку в отдельный файл
(cherry picked from commit f2b97cd9d82375543405e95b854fcb335c40ca4a)
